RES-2499: Move meta_continual_learning code from frameworks to standalone package
