+++ nupic/research/frameworks/vernon/experiments/__init__.py
@@ -20,6 +20,5 @@
 # ----------------------------------------------------------------------
 
 from .cl_experiment import *
-from .meta_cl_experiment import *
 from .self_supervised_experiment import *
 from .supervised_experiment import *
