+++ packages/meta_continual_learning/nupic/research/frameworks/meta_continual_learning/experiments/meta_cl_experiment.py
@@ -26,7 +26,7 @@ import numpy as np
 import torch
 from torch.utils.data import DataLoader
 
-from nupic.research.frameworks.continual_learning.maml_utils import clone_model
+from nupic.research.frameworks.meta_continual_learning.maml_utils import clone_model
 from nupic.research.frameworks.pytorch.dataset_utils.samplers import TaskRandomSampler
 from nupic.research.frameworks.pytorch.model_utils import (
     filter_params,
