+++ projects/meta_cl/experiments/oml_variants.py
@@ -23,9 +23,13 @@ from copy import deepcopy
 
 from experiment_classes import IIDTrainingOMLExperiment, OMLExperiment
 from networks import KWinnerOMLNetwork
+from nupic.research.frameworks.meta_continual_learning import mixins as metacl_mixins
+from nupic.research.frameworks.meta_continual_learning.experiments import (
+    MetaContinualLearningExperiment,
+)
+from nupic.research.frameworks.meta_continual_learning.models import OMLNetwork
 from nupic.research.frameworks.pytorch.datasets import omniglot
-from nupic.research.frameworks.pytorch.models import OMLNetwork
-from nupic.research.frameworks.vernon import MetaContinualLearningExperiment, mixins
+from nupic.research.frameworks.vernon import mixins
 
 from .anml_replicate import metacl_anml_replicate
 from .oml_replicate import metacl_oml_replicate, oml_datasplit_without_norm
@@ -183,7 +187,7 @@ oml_datasplit_without_norm_2000.update(
 )
 
 
-class KWinnerOMLExperiment(mixins.OnlineMetaLearning,
+class KWinnerOMLExperiment(metacl_mixins.OnlineMetaLearning,
                            mixins.UpdateBoostStrength,
                            MetaContinualLearningExperiment):
     pass
