+++ setup.cfg
@@ -160,14 +160,6 @@ include = nupic.research.*
 [bdist_wheel]
 universal=0
 
-[aliases]
-test=pytest
-
-[tool:pytest]
-testpaths=tests
-filterwarnings =
-    ignore:.*:DeprecationWarning
-
 [flake8]
 ; Select all
 select = A,B,C,D,E,F,G,H,I,J,K,L,M,N,O,P,Q,R,S,T,U,V,W,X,Y,Z
