+++ tests/unit/frameworks/pytorch/filter_params_test.py
@@ -24,7 +24,7 @@ import unittest
 import torch
 
 from nupic.research.frameworks.pytorch.model_utils import filter_params
-from nupic.research.frameworks.pytorch.models import OMLNetwork, resnets
+from nupic.research.frameworks.pytorch.models import resnets
 
 
 class FilterParamsTest(unittest.TestCase):
@@ -37,37 +37,37 @@ class FilterParamsTest(unittest.TestCase):
         Test use of `include_names`.
         """
 
-        oml = OMLNetwork(num_classes=10)
-        named_params = filter_params(oml, include_names=["adaptation.0.weight"])
-        self.assertTrue(len(named_params) == 1)
-        self.assertTrue("adaptation.0.weight" in named_params)
-        self.assertTrue(named_params["adaptation.0.weight"].shape == (10, 2304))
+        resnet = resnets.resnet50(num_classes=10)
+        named_params = filter_params(resnet, include_names=["classifier.weight"])
+        self.assertEqual(len(named_params), 1)
+        self.assertIn("classifier.weight", named_params)
+        self.assertEqual(named_params["classifier.weight"].shape, (10, 2048))
 
     def test_name_not_included(self):
         """
         Test the case when a param name does not exist in the network.
         """
-
-        oml = OMLNetwork(num_classes=10)
-
-        named_params = filter_params(oml, include_names=["adaptation.0"])
-        self.assertTrue(len(named_params) == 0)
+        resnet = resnets.resnet50(num_classes=10)
+        named_params = filter_params(resnet, include_names=["classifier"])
+        self.assertEqual(len(named_params), 0)
 
     def test_get_conv_modules_by_pattern_and_type(self):
         """
         Ensure `include_patterns` and `include_modules` yields the same result
         when they are meant to identify the same params.
         """
-        oml = OMLNetwork(num_classes=10)
-        named_params1 = filter_params(oml, include_patterns=["representation.*"])
-        self.assertTrue(len(named_params1) == 12)  # 6 convs with a weight and bias each
+        resnet = resnets.resnet50(num_classes=10)
+        named_params1 = filter_params(
+            resnet, include_patterns=["classifier"])
+        self.assertEqual(len(named_params1), 2)  # linear weight + bias
 
-        named_params2 = filter_params(oml, include_modules=[torch.nn.Conv2d])
-        self.assertTrue(len(named_params2) == 12)
+        named_params2 = filter_params(
+            resnet, include_modules=[torch.nn.Linear])
+        self.assertEqual(len(named_params2), 2)
 
         names1 = list(named_params1.keys())
         names2 = list(named_params2.keys())
-        self.assertTrue(names1 == names2)
+        self.assertEqual(names1, names2)
 
     def test_filter_out_resnet_linear_params(self):
         """
@@ -75,9 +75,9 @@ class FilterParamsTest(unittest.TestCase):
         """
         resnet = resnets.resnet50(num_classes=10)
         named_params = filter_params(resnet, include_modules=[torch.nn.Linear])
-        self.assertTrue(len(named_params) == 2)
-        self.assertTrue("classifier.weight" in named_params)
-        self.assertTrue("classifier.bias" in named_params)
+        self.assertEqual(len(named_params), 2)
+        self.assertIn("classifier.weight", named_params)
+        self.assertIn("classifier.bias", named_params)
 
 
 if __name__ == "__main__":
